+++ database/migrations/2020_02_24_134129_create_tags_table.php
@@ -17,7 +17,7 @@ class CreateTagsTable extends Migration
             $table->bigIncrements('id');
             $table->string('name');
             $table->string('slug');
-            $table->integer('status');
+            $table->tinyInteger('status');
             $table->timestamps();
         });
     }
